+++ lib/core/data/size.hpp
@@ -8,14 +8,14 @@
 
     public:
 
-      unsigned long int horz;
-      unsigned long int vert;
-      unsigned long int depth;
+      float horz;
+      float vert;
+      float depth;
 
       Size(
-        unsigned long int width=0,
-        unsigned long int height=0,
-        unsigned long int depth=0
+        float width=0.0,
+        float height=0.0,
+        float depth=0.0
       ) {
 
         this->horz = width;;
